add bet limit

+++ src/commands/fun/blackjack.js
@@ -94,6 +94,7 @@ module.exports = {
         if (bet && user.balance < bet) {
             return message.reply(`❌ You don't have enough money! Balance: **${user.balance}**`);
         }
+        if (bet > 250000) return message.reply('❌ The maximum bet is **250,000** coins!');
         if (bet) db.removeBalance(user.id, bet);
 
         const playerHand = [drawCard(), drawCard()];
